refactor: 삭제 dialog - AppView zstack에 뜨도록 변경
